Remove outdated section on programmatic access (#230)
This section makes it seem like programmatic access is still not
available, when in fact it is. Instead of updating the section, I opted
for removing it because the same topic is already covered under "next
steps".

Guide the user towards using the huggingface_hub library first before
explaining how to use git directly. Changed the wording a bit to make it
fit.