Update installation of spacy models from hub spacy.md (#1422)
The update in the way to install spacy models from the hub comes from how pip parse the URL to check pacage name and version. As it detect's "any" as version, it can raise an error. Adding the name of the package alongside the URL solve this. 
More info in this GitHub issue: https://github.com/explosion/spaCy/issues/13599